+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Implement `Tag` component
 -   Documentation on how to handle library dependencies
+-   Implement `Avatar` component
 
 ### Changed